Add thumbnail deletion when deleting video recordings

+++ opentakserver/blueprints/ots_api/video_api.py
@@ -65,7 +65,15 @@ def download_recording():
             filename = pathlib.Path(recording.segment_path)
             return send_from_directory(filename.parent, filename.name)
         elif request.method == 'DELETE':
-            os.remove(recording.segment_path)
+            # Delete video file
+            if os.path.exists(recording.segment_path):
+                os.remove(recording.segment_path)
+            
+            # Delete thumbnail if exists
+            if recording.thumbnail and os.path.exists(recording.thumbnail):
+                os.remove(recording.thumbnail)
+            
+            # Delete from database
             db.session.delete(recording)
             db.session.commit()
             return jsonify({'success': True})
